[~] Fix warnings
