+++ package.json
@@ -45,13 +45,6 @@
   "dependencies": {
     "@hkube/consts": "^1.0.21",
     "@reduxjs/toolkit": "^1.3.2",
-    "@vx/gradient": "^0.0.195",
-    "@vx/group": "^0.0.195",
-    "@vx/hierarchy": "^0.0.195",
-    "@vx/network": "^0.0.195",
-    "@vx/shape": "^0.0.195",
-    "d3-hierarchy": "^1.1.9",
-    "d3-shape": "^1.3.7",
     "framer-motion": "^1.10.3",
     "gatsby": "^2.20.9",
     "humanize-duration": "^3.22.0",
@@ -65,10 +58,8 @@
     "react-helmet": "^5.2.1",
     "react-moment": "^0.9.7",
     "react-redux": "^7.2.0",
-    "react-tooltip": "^4.1.3",
     "react-virtualized-auto-sizer": "^1.0.2",
     "react-window": "^1.8.5",
-    "redux-logger": "^3.0.6",
     "reusable": "^1.0.0",
     "socket.io-client": "^2.3.0",
     "styled-components": "^5.0.0",
@@ -79,7 +70,7 @@
   "devDependencies": {
     "@dennisvash/eslint-config": "^3.0.0",
     "@dennisvash/prettier-config": "^2.0.1",
-    "@storybook/react": "^5.3.8",
+    "@storybook/react": "^5.3.18",
     "babel-eslint": "^10.0.3",
     "babel-loader": "^8.1.0",
     "babel-plugin-styled-components": "^1.10.7",
@@ -97,7 +88,7 @@
     "gatsby-plugin-resolve-src": "^2.0.0",
     "gatsby-plugin-styled-components": "^3.2.1",
     "husky": "4.2.3",
-    "lint-staged": "^10.1.0",
+    "lint-staged": "^10.1.1",
     "prettier": "^2.0.2",
     "pretty-quick": "^2.0.1",
     "prop-types": "^15.7.2",
