+++ src/components/molecules/Job/JobGraph/JobGraph.stories.js
@@ -1,8 +1,8 @@
 import { GRAPH, SB_LABELS } from '@constants';
-import { useGraph } from '@hooks';
+import { useGraph, useJobs } from '@hooks';
 import { mixins } from '@styles';
-import { JobSelectHelper } from '@utils';
-import React, { useMemo, useReducer } from 'react';
+import { NOOP } from '@utils';
+import React, { useCallback, useEffect, useMemo, useReducer } from 'react';
 import styled from 'styled-components';
 import tw from 'twin.macro';
 import JobGraph from './JobGraph.react';
@@ -18,6 +18,35 @@ const Flex = styled.div`
 const Button = styled.button`
   ${tw`border-black border`}
 `;
+
+const Container = styled.div`
+  ${tw`border-black p-2 border-2 m-2`}
+`;
+
+const JobSelectHelper = () => {
+  const { select, list, selected } = useJobs();
+
+  const onChange = useCallback(({ target: { value } }) => select(value), [select]);
+
+  useEffect(() => {
+    if (selected === null && list.length !== 0) {
+      select(list[0]?.jobId);
+    }
+  }, [list, select, selected]);
+
+  return (
+    <Container>
+      <select onChange={onChange} onBlur={NOOP}>
+        {list.map(({ jobId }) => (
+          <option key={jobId} value={jobId}>
+            {jobId}
+          </option>
+        ))}
+      </select>
+      <div>Selected: {selected || `None`}</div>
+    </Container>
+  );
+};
 
 export default {
   title: `${SB_LABELS.MOLECULES}Jobs/Job Graph`,
