+++ src/state/store/middleware.js
@@ -1,7 +1,4 @@
 import { getDefaultMiddleware } from '@reduxjs/toolkit';
-// import { createLogger } from 'redux-logger';
-
-// const logger = createLogger({ collapsed: true });
 
 const middleware = [
   ...getDefaultMiddleware({
