+++ src/utils/index.js
@@ -2,5 +2,4 @@ export { IS_WINDOW_DEFINED, NOOP } from './common';
 export { areEqualGraphs, formatEdge, formatNode } from './graph';
 export { getBooleanLSItem, getLsItem, getLsObjectItem, setLsItem } from './localStorage';
 export { calc, trans } from './parallax';
-export { JobSelectHelper } from './storybook';
 export { sorter, stringify, toBoolean, toUpperCaseFirstLetter } from './string';
